Centraliza dados e ajusta layout para o portal de serviços
