+++ src/lib/data.ts
@@ -1,0 +1,395 @@
+const issItems = [
+  { text: "Emitir Guia ISS/TLFF/TLA/TRIFS", href: "#" },
+  {
+    text: "Emitir Guia ISS/TLFF/TLA/TRIFS Autônomo",
+    href: "#",
+  },
+  {
+    text: "Nota Fiscal de Serviços",
+    href: "#",
+  },
+  {
+    text: "Nota Fiscal de Serviços Avulsa",
+    href: "#",
+  },
+  {
+    text: "Emitir Cartão de Inscrição de Empresa/Autônomo",
+    href: "#",
+  },
+  {
+    text: "Certidão Conjunta Negativa de Débitos Municipais e da Dívida Ativa do Município",
+    href: "#",
+  },
+  {
+    text: "Certidão Conjunta Positiva de Débitos Municipais e da Dívida Ativa do Município",
+    href: "#",
+  },
+  { text: "Reemitir Certidão/Declaração", href: "#" },
+  { text: "Validar Certidão/Declaração", href: "#" },
+  { text: "Certidão de Baixa", href: "#" },
+  { text: "Certidão de Estimativa", href: "#" },
+  { text: "Consultar Alvará de Funcionamento", href: "#" },
+  {
+    text: "Declaração de Inexistência de Cadastro Mercantil",
+    href: "#",
+  },
+  {
+    text: "Portal de Parcelamento e Negociação",
+    href: "#",
+  },
+  { text: "Legislação", href: "#" },
+  {
+    text: "Notificação de Instrução Normativa",
+    href: "#",
+  },
+];
+
+const iptuItems = [
+  { text: "Certidão Negativa de Débitos de IPTU", href: "#" },
+  {
+    text: "Certidão Positiva com Efeitos de Negativa IPTU",
+    href: "#",
+  },
+  {
+    text:
+      "Emitir Guia IPTU/TCRD-Taxa de Lixo/COSIP -\n" +
+      "Taxa de Iluminação Pública",
+    href: "#",
+  },
+  {
+    text: "Emitir Memória de Cálculo do Imóvel",
+    href: "#",
+  },
+  {
+    text: "Declaração de Inexistência de Imóvel",
+    href: "#",
+  },
+  {
+    text: "Reemitir Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Validar Certidão/Declaração",
+    href: "#",
+  },
+  { text: "Portal do IPTU", href: "#" },
+  { text: "Portal de Negociação", href: "#" },
+];
+
+const itbiItems = [
+  { text: "Emitir Guia ITBI-e", href: "#" },
+  {
+    text: "Emitir Guia SIAT",
+    href: "#",
+  },
+  {
+    text: "Sistema ITBI-e",
+    href: "#",
+  },
+  {
+    text: "Certidão Negativa de Transferência Imobiliária",
+    href: "#",
+  },
+  {
+    text: "Reemitir Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Validar Certidão/Declaração",
+    href: "#",
+  },
+];
+
+const taxaItems = [
+  { text: "Emitir Guia ISS/TLFF/TLA/TRIFS", href: "#" },
+  {
+    text: "Emitir Guia ISS/TLFF/TLA/TRIFS Autonômo",
+    href: "#",
+  },
+  {
+    text: "Emitir Cartão de Inscrição de Empresa/Autônomo",
+    href: "#",
+  },
+  {
+    text:
+      "Certidão Conjunta Negativa de Débitos\n" +
+      "Municipais e da Dívida Ativa do Município",
+    href: "#",
+  },
+  {
+    text:
+      "Certidão Conjunta Positiva de Débitos\n" +
+      "Municipais e da Dívida Ativa do Município",
+    href: "#",
+  },
+  {
+    text: "Reemitir Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Consultar Alvará de Funcionamento",
+    href: "#",
+  },
+  {
+    text: "Declaração de Inexistência de Cadastro Mercantil",
+    href: "#",
+  },
+  {
+    text: "Portal de Parcelamento e Negociação",
+    href: "#",
+  },
+  {
+    text: "Sistema de Licenciamento - SLIC",
+    href: "#",
+  },
+];
+
+const imovelItens = [
+  { text: "Certidão Negativa de Débitos de IPTU", href: "#" },
+  {
+    text: "Certidão Positiva com Efeitos de Negativa IPTU",
+    href: "#",
+  },
+  {
+    text:
+      "Emitir Guia IPTU/TCRD-Taxa de Lixo/COSIP -\n" +
+      "Taxa de Iluminação Pública",
+    href: "#",
+  },
+  {
+    text: "Emitir Memória de Cálculo do Imóvel",
+    href: "#",
+  },
+  {
+    text: "Emitir Guia ITBI-e",
+    href: "#",
+  },
+  {
+    text: "Emitir Guia SIAT",
+    href: "#",
+  },
+  {
+    text: "Sistema ITBI-e",
+    href: "#",
+  },
+  {
+    text: "Declaração de Inexistência de Imóvel",
+    href: "#",
+  },
+  {
+    text: "Reemitir Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Validar Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Portal do IPTU",
+    href: "#",
+  },
+  {
+    text: "Portal de Negociação",
+    href: "#",
+  },
+  {
+    text: "Sistema Coleta de Lixo",
+    href: "#",
+  },
+  {
+    text: "Sistema de Coleta de Poda",
+    href: "#",
+  },
+  {
+    text: "Certidão de Número",
+    href: "#",
+  },
+  {
+    text: "Declaração de Rua",
+    href: "#",
+  },
+  {
+    text: "Construa Fácil",
+    href: "#",
+  },
+];
+
+const empresaItens = [
+  { text: "Emitir Guia ISS/TLFF/TLA/TRIFS", href: "#" },
+  {
+    text: "Emitir Guia ISS/TLFF/TLA/TRIFS Autonômo",
+    href: "#",
+  },
+  {
+    text: "Nota Fiscal de Serviços",
+    href: "#",
+  },
+  {
+    text: "Nota Fiscal de Serviços Avulsa",
+    href: "#",
+  },
+  {
+    text: "Emitir Cartão de Inscrição de Empresa/Autônomo",
+    href: "#",
+  },
+  {
+    text:
+      "Certidão Conjunta Negativa de Débitos\n" +
+      "Municipais e da Dívida Ativa do Município",
+    href: "#",
+  },
+  {
+    text:
+      "Certidão Conjunta Negativa de Débitos\n" +
+      "Municipais e da Dívida Ativa do Município",
+    href: "#",
+  },
+  {
+    text: "Reemitir Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Reemitir Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Validar Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Certidão de Baixa",
+    href: "#",
+  },
+  {
+    text: "Certidão de Estimativa",
+    href: "#",
+  },
+  {
+    text: "Consultar Alvará de Funcionamento",
+    href: "#",
+  },
+  {
+    text: "Declaração de Inexistência de Cadastro Mercantil",
+    href: "#",
+  },
+  {
+    text: "Portal de Parcelamento e Negociação",
+    href: "#",
+  },
+  {
+    text: "Notificação de Instrução Normativa",
+    href: "#",
+  },
+  {
+    text: "Sistema de Licenciamento - SLIC",
+    href: "#",
+  },
+  {
+    text: "Piauí Digital",
+    href: "#",
+  },
+  {
+    text: "DESIF",
+    href: "#",
+  },
+  {
+    text: "Sistema de Poda",
+    href: "#",
+  },
+  {
+    text: "Sistema Coleta de Lixo",
+    href: "#",
+  },
+  {
+    text: "Domicílio Tributário Eletrônico",
+    href: "#",
+  },
+];
+
+const cidadaoItems = [
+  { text: "Certidão Negativa de Débitos de IPTU", href: "#" },
+  {
+    text: "Certidão Positiva com Efeitos de Negativa IPTU",
+    href: "#",
+  },
+  {
+    text:
+      "Emitir Guia IPTU/TCRD-Taxa de Lixo/COSIP -\n" +
+      "Taxa de Iluminação Pública",
+    href: "#",
+  },
+  {
+    text: "Emitir Memória de Cálculo do Imóvel",
+    href: "#",
+  },
+  {
+    text: "Emitir Guia ITBI-e",
+    href: "#",
+  },
+  {
+    text: "Emitir Guia SIAT",
+    href: "#",
+  },
+  {
+    text: "Sistema ITBI-e",
+    href: "#",
+  },
+  {
+    text: "Declaração de Inexistência de Imóvel",
+    href: "#",
+  },
+  {
+    text: "Declaração de Inexistência de Cadastro Mercantil",
+    href: "#",
+  },
+  {
+    text: "Reemitir Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Validar Certidão/Declaração",
+    href: "#",
+  },
+  {
+    text: "Portal do IPTU",
+    href: "#",
+  },
+  {
+    text: "Portal de Negociação",
+    href: "#",
+  },
+  {
+    text: "Sistema Coleta de Lixo",
+    href: "#",
+  },
+  {
+    text: "Sistema de Coleta de Poda",
+    href: "#",
+  },
+  {
+    text: "Certidão de Número",
+    href: "#",
+  },
+  {
+    text: "Declaração de Rua",
+    href: "#",
+  },
+  {
+    text: "Construa Fácil",
+    href: "#",
+  },
+  {
+    text: "Cadastro como Tomador da Nota",
+    href: "#",
+  },
+];
+
+export {
+  issItems,
+  iptuItems,
+  itbiItems,
+  taxaItems,
+  imovelItens,
+  empresaItens,
+  cidadaoItems,
+};
